backend_service/requirements.txt
Flask==2.2.2
PyJWT==2.6.0
werkzeug==2.2.2
